Use platform specific imports for Rollbar.js (#172)
* rename Rollbar.js file to Client.js

* update ts type definitions

* use platform specific import for Rollbar.js
