+++ lib/Rollbar.native.js
@@ -1,0 +1,2 @@
+Object.defineProperty(exports,"__esModule",{value:true});exports.default=void 0;var _rollbar=_interopRequireDefault(require("rollbar/src/react-native/rollbar"));function _interopRequireDefault(obj){return obj&&obj.__esModule?obj:{default:obj};}var _default=_rollbar.default;exports.default=_default;
+//# sourceMappingURL=Rollbar.native.js.map
