+++ lib/merge.js
@@ -1,0 +1,2 @@
+'use strict';Object.defineProperty(exports,"__esModule",{value:true});exports.merge=merge;var hasOwn=Object.prototype.hasOwnProperty;var toStr=Object.prototype.toString;var isPlainObject=function isPlainObject(obj){if(!obj||toStr.call(obj)!=='[object Object]'){return false;}var hasOwnConstructor=hasOwn.call(obj,'constructor');var hasIsPrototypeOf=obj.constructor&&obj.constructor.prototype&&hasOwn.call(obj.constructor.prototype,'isPrototypeOf');if(obj.constructor&&!hasOwnConstructor&&!hasIsPrototypeOf){return false;}var key;for(key in obj){}return typeof key==='undefined'||hasOwn.call(obj,key);};function merge(){var i,src,copy,clone,name,result={},current=null,length=arguments.length;for(i=0;i<length;i++){current=arguments[i];if(current==null){continue;}for(name in current){src=result[name];copy=current[name];if(result!==copy){if(copy&&isPlainObject(copy)){clone=src&&isPlainObject(src)?src:{};result[name]=merge(clone,copy);}else if(typeof copy!=='undefined'){result[name]=copy;}}}}return result;}
+//# sourceMappingURL=merge.js.map
