+++ src/Rollbar.native.js
@@ -1,0 +1,3 @@
+import Rollbar from 'rollbar/src/react-native/rollbar';
+
+export default Rollbar;
